Bugs fixed 1) no push for pull now 2) Integer comparision epic fail

+++ src/main/java/dibd/daemon/command/OverCommand.java
@@ -247,7 +247,7 @@ public class OverCommand implements Command {
         overview.append(escapeString(art.getMessageId()))
         .append('\t');
         //6) thread-Id
-        if(art.getThread_id()!=art.getId()) //if replay
+        if(art.getThread_id().intValue() != art.getId().intValue()) //if replay
         	overview.append(StorageManager.current().getMessageId(art.getThread_id()));
         //overview.append('\t');
